refactor fiter memory
