+++ src/components/ScreeningList/Filter/index.jsx
@@ -15,17 +15,8 @@ import Tooltip from '@components/Tooltip';
 import Button from '@components/Button';
 import Icon from '@components/Icon';
 import Badge from '@components/Badge';
-import Menu from '@components/Menu';
-import Dropdown from '@components/Dropdown';
-import notification from '@components/notification';
 import LoadBox from '@components/LoadBox';
-import {
-  FILTER_PRIVATE_STORE_ID,
-  FILTER_PRIVATE_MEMORY_TYPE,
-  FILTER_PUBLIC_STORE_ID,
-  FILTER_PUBLIC_MEMORY_TYPE
-} from '@utils/memory';
-import SaveFilterModal from './components/SaveFilterModal';
+import FilterMemory from './components/FilterMemory';
 
 import { Box, SearchBox } from './Filter.style';
 import './index.css';
@@ -52,7 +43,7 @@ export default function Filter({
   publicFilters
 }) {
   const [open, setOpen] = useState(false);
-  const [saveFilterOpen, setSaveFilterOpen] = useState(false);
+
   const [date, setDate] = useState([moment(match.params.startDate), null]);
   const { t, i18n } = useTranslation();
 
@@ -92,11 +83,6 @@ export default function Filter({
   );
 
   useEffect(() => {
-    fetchMemory(FILTER_PRIVATE_STORE_ID, `${FILTER_PRIVATE_MEMORY_TYPE}_${account.userId}`);
-    fetchMemory(FILTER_PUBLIC_STORE_ID, FILTER_PUBLIC_MEMORY_TYPE);
-  }, [account.userId, fetchMemory]);
-
-  useEffect(() => {
     if (!isEmpty(segments.error)) {
       message.error(segments.error.message);
     }
@@ -127,84 +113,13 @@ export default function Filter({
     };
   }, [updateStatus]);
 
-  const filterSubmenu = (list, type) => {
-    const title = t(`screeningList.${type}Filter`);
-
-    if (isEmpty(list) || isEmpty(list[0].value)) {
-      return (
-        <Menu.SubMenu title={title}>
-          <Menu.Item>
-            {t('screeningList.noFilter')}
-            <br /> {t('screeningList.noFilterHint')}
-          </Menu.Item>
-        </Menu.SubMenu>
-      );
+  const loadFilter = filterData => {
+    fetchPrescriptionsList(getParams({ ...filterData, idDept: filterData.idDepartment }));
+    if (!isEmpty(filterData.idDrug)) {
+      searchDrugs(filterData.idSegment, { idDrug: filterData.idDrug });
     }
-
-    return (
-      <Menu.SubMenu title={title}>
-        {list[0].value.map((item, index) => (
-          <Menu.Item key={index} onClick={() => loadFilterAction(item.data)}>
-            {item.name}
-          </Menu.Item>
-        ))}
-      </Menu.SubMenu>
-    );
+    setOpen(false);
   };
-
-  const removeFilterSubmenu = (list, type) => {
-    const title = t(`screeningList.${type}FilterRemove`);
-
-    if (list && list.length && !isEmpty(list[0].value)) {
-      return (
-        <Menu.SubMenu title={title}>
-          {list[0].value.map((item, index) => (
-            <Menu.Item
-              key={index}
-              onClick={() => removeFilterAction(index, type)}
-              style={{ color: '#ff4d4f' }}
-            >
-              <Icon type="delete" style={{ fontSize: 16, color: '#ff4d4f' }} />
-              {item.name}
-            </Menu.Item>
-          ))}
-        </Menu.SubMenu>
-      );
-    }
-
-    return null;
-  };
-
-  const filterMenu = () => (
-    <Menu forceSubMenuRender={true}>
-      <Menu.Item className="gtm-btn-menu-filter-save" onClick={() => setSaveFilterOpen(true)}>
-        {t('screeningList.saveFilter')}
-      </Menu.Item>
-
-      {!isEmpty(savedFilters) && (
-        <Menu.SubMenu title="Filtros locais">
-          {savedFilters.map((item, index) => (
-            <Menu.Item key={index} onClick={() => loadFilterAction(item.data)}>
-              <Tooltip title="Os filtros locais serão desativados. Utilize os filtros privados ou públicos.">
-                {item.name}
-              </Tooltip>
-            </Menu.Item>
-          ))}
-          <Menu.Item key={999} style={{ color: '#ff4d4f' }}>
-            Os filtros locais serão desativados.
-            <br /> Utilize os filtros privados ou públicos.
-          </Menu.Item>
-        </Menu.SubMenu>
-      )}
-      {filterSubmenu(privateFilters, 'private')}
-      {filterSubmenu(publicFilters, 'public')}
-
-      <Menu.Divider />
-
-      {removeFilterSubmenu(privateFilters, 'private')}
-      {removeFilterSubmenu(publicFilters, 'public')}
-    </Menu>
-  );
 
   const onDepartmentChange = idDept => {
     setScreeningListFilter({ idDepartment: idDept });
@@ -279,70 +194,6 @@ export default function Filter({
     });
 
     return count;
-  };
-
-  const saveFilterAction = (filterName, filterType) => {
-    if (filterType === 'public') {
-      const hasFilter = publicFilters && publicFilters.length;
-      const filters = hasFilter ? [...publicFilters[0].value] : [];
-      filters.push({
-        name: filterName,
-        data: filter
-      });
-
-      saveMemory(FILTER_PUBLIC_STORE_ID, {
-        id: hasFilter ? publicFilters[0].key : null,
-        type: FILTER_PUBLIC_MEMORY_TYPE,
-        value: filters
-      });
-    } else {
-      const hasFilter = privateFilters && privateFilters.length;
-      const filters = hasFilter ? [...privateFilters[0].value] : [];
-      filters.push({
-        name: filterName,
-        data: filter
-      });
-
-      saveMemory(FILTER_PRIVATE_STORE_ID, {
-        id: hasFilter ? privateFilters[0].key : null,
-        type: `${FILTER_PRIVATE_MEMORY_TYPE}_${account.userId}`,
-        value: filters
-      });
-    }
-
-    notification.success({ message: 'Uhu! Filtro salvo com sucesso!' });
-  };
-
-  const removeFilterAction = (index, type) => {
-    const storeElement = type === 'public' ? publicFilters[0] : privateFilters[0];
-    const filters = [...storeElement.value];
-
-    filters.splice(index, 1);
-
-    if (type === 'public') {
-      saveMemory(FILTER_PUBLIC_STORE_ID, {
-        id: storeElement.key,
-        type: FILTER_PUBLIC_MEMORY_TYPE,
-        value: filters
-      });
-    } else {
-      saveMemory(FILTER_PRIVATE_STORE_ID, {
-        id: storeElement.key,
-        type: `${FILTER_PRIVATE_MEMORY_TYPE}_${account.userId}`,
-        value: filters
-      });
-    }
-
-    notification.success({ message: 'Filtro removido com sucesso!' });
-  };
-
-  const loadFilterAction = filterData => {
-    setScreeningListFilter(filterData);
-    fetchPrescriptionsList(getParams({ ...filterData, idDept: filterData.idDepartment }));
-    if (!isEmpty(filterData.idDrug)) {
-      searchDrugs(filterData.idSegment, { idDrug: filterData.idDrug });
-    }
-    setOpen(false);
   };
 
   const searchDrugsAutocomplete = debounce(value => {
@@ -424,14 +275,17 @@ export default function Filter({
                 loading={isFetchingPrescription}
               />
             </Tooltip>
-            <Dropdown overlay={filterMenu()}>
-              <Button
-                className="gtm-btn-filter"
-                shape="circle"
-                icon="filter"
-                style={{ marginTop: '11px', marginLeft: '5px' }}
-              />
-            </Dropdown>
+            <FilterMemory
+              fetchMemory={fetchMemory}
+              account={account}
+              savedFilters={savedFilters}
+              publicFilters={publicFilters}
+              privateFilters={privateFilters}
+              saveMemory={saveMemory}
+              filter={filter}
+              setScreeningListFilter={setScreeningListFilter}
+              loadFilter={loadFilter}
+            />
           </div>
         </Col>
       </Row>
@@ -547,12 +401,6 @@ export default function Filter({
           </Row>
         </Col>
       </Row>
-
-      <SaveFilterModal
-        setSaveFilterOpen={setSaveFilterOpen}
-        saveFilterAction={saveFilterAction}
-        open={saveFilterOpen}
-      />
     </SearchBox>
   );
 }
